+++ Frends.SFTP.RenameFile/Frends.SFTP.RenameFile.Tests/ConnectivityTests.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.IO;
+using System.Threading.Tasks;
 using NUnit.Framework;
 using Frends.SFTP.RenameFile.Enums;
 
@@ -9,26 +10,36 @@ namespace Frends.SFTP.RenameFile.Tests;
 public class ConnectivityTests : RenameFileTestBase
 {
     [Test]
-    public void RenameFile_TestWithPrivateKeyFileRsa()
+    public async Task RenameFile_TestWithPrivateKeyFileRsa()
     {
         _connection.Authentication = AuthenticationType.UsernamePasswordPrivateKeyFile;
-        _connection.PrivateKeyFilePassphrase = "passphrase";
+        _connection.PrivateKeyPassphrase = "passphrase";
         _connection.PrivateKeyFile = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "../../../Volumes/ssh_host_rsa_key");
 
-        var result = SFTP.RenameFile(_input, _connection);
+        var result = await SFTP.RenameFile(_input, _connection, default);
         Assert.IsTrue(Helpers.DestinationFileExists(result.Path));
     }
 
     [Test]
-    public void RenameFile_TestWithPrivateKeyFileRsaFromString()
+    public async Task RenameFile_TestWithPrivateKeyFileRsaFromString()
     {
         var key = File.ReadAllText(Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "../../../Volumes/ssh_host_rsa_key"));
 
         _connection.Authentication = AuthenticationType.UsernamePasswordPrivateKeyString;
-        _connection.PrivateKeyFilePassphrase = "passphrase";
+        _connection.PrivateKeyPassphrase = "passphrase";
         _connection.PrivateKeyString = key;
 
-        var result = SFTP.RenameFile(_input, _connection);
+        var result = await SFTP.RenameFile(_input, _connection, default);
+        Assert.IsTrue(Helpers.DestinationFileExists(result.Path));
+    }
+
+    [Test]
+    public async Task RenameFile_TestWithKeyboardInteractiveAuthentication()
+    {
+        _connection.Authentication = AuthenticationType.UsernamePassword;
+        _connection.UseKeyboardInteractiveAuthentication = true;
+
+        var result = await SFTP.RenameFile(_input, _connection, default);
         Assert.IsTrue(Helpers.DestinationFileExists(result.Path));
     }
 }