+++ Frends.SFTP.RenameFile/Frends.SFTP.RenameFile.Tests/Lib/Helpers.cs
@@ -39,61 +39,40 @@ internal static class Helpers
     internal static void GenerateDummyFile(string path)
     {
         var content = "This is a test file.";
-        using (var client = new SftpClient(_dockerAddress, 2222, _dockerUsername, _dockerPassword))
-        {
-            client.Connect();
-            client.WriteAllText(path, content);
-            client.Disconnect();
-        }
+        using var client = new SftpClient(_dockerAddress, 2222, _dockerUsername, _dockerPassword);
+        client.Connect();
+        client.WriteAllText(path, content);
+        client.Disconnect();
     }
 
-    internal static Tuple<byte[], byte[]> GetServerFingerPrintAndHostKey()
+    internal static Tuple<string, string, byte[]> GetServerFingerPrintsAndHostKey()
     {
-        Tuple<byte[], byte[]> result = null;
-        using (var client = new SftpClient(_dockerAddress, 2222, _dockerUsername, _dockerPassword))
-        {
-            client.ConnectionInfo.HostKeyAlgorithms.Clear();
-            client.ConnectionInfo.HostKeyAlgorithms.Add("ssh-rsa", (data) => { return new KeyHostAlgorithm("ssh-rsa", new RsaKey(), data); });
+        Tuple<string, string, byte[]> result = null;
+        using var client = new SftpClient(_dockerAddress, 2222, _dockerUsername, _dockerPassword);
+        client.ConnectionInfo.HostKeyAlgorithms.Clear();
+        client.ConnectionInfo.HostKeyAlgorithms.Add("ssh-rsa", (data) => { return new KeyHostAlgorithm("ssh-rsa", new RsaKey(), data); });
 
-            client.HostKeyReceived += delegate (object sender, HostKeyEventArgs e)
-            {
-                result = new Tuple<byte[], byte[]>(e.FingerPrint, e.HostKey);
-                e.CanTrust = true;
-            };
-            client.Connect();
-            client.Disconnect();
-        }
+        client.HostKeyReceived += delegate (object sender, HostKeyEventArgs e)
+        {
+            result = new Tuple<string, string, byte[]>(e.FingerPrintMD5, e.FingerPrintSHA256, e.HostKey);
+            e.CanTrust = true;
+        };
+        client.Connect();
+        client.Disconnect();
         return result;
-    }
-
-    internal static string ConvertToMD5Hex(byte[] fingerPrint)
-    {
-        return BitConverter.ToString(fingerPrint).Replace("-", ":");
-    }
-
-    internal static string ConvertToSHA256Hash(byte[] hostKey)
-    {
-        var fingerprint = "";
-        using (SHA256 mySHA256 = SHA256.Create())
-        {
-            fingerprint = Convert.ToBase64String(mySHA256.ComputeHash(hostKey));
-        }
-        return fingerprint;
     }
 
     internal static string ConvertToSHA256Hex(byte[] hostKey)
     {
         var fingerprint = "";
-        using (SHA256 mySHA256 = SHA256.Create())
-        {
-            fingerprint = ToHex(mySHA256.ComputeHash(hostKey));
-        }
+        using var mySHA256 = SHA256.Create();
+        fingerprint = ToHex(mySHA256.ComputeHash(hostKey));
         return fingerprint;
     }
 
     internal static string ToHex(byte[] bytes)
     {
-        StringBuilder result = new StringBuilder(bytes.Length * 2);
+        var result = new StringBuilder(bytes.Length * 2);
         for (int i = 0; i < bytes.Length; i++)
             result.Append(bytes[i].ToString("x2"));
         return result.ToString();
@@ -103,10 +82,8 @@ internal static class Helpers
     {
         var files = client.ListDirectory(dir).Where(f => (f.Name != ".") && (f.Name != ".."));
         foreach (var file in files)
-        {
             if (!file.IsDirectory)
                 client.DeleteFile(file.FullName);
-        }
     }
 
     internal static bool DestinationFileExists(string dest)
@@ -121,7 +98,6 @@ internal static class Helpers
     internal static SshKeyGenerator.SshKeyGenerator GenerateDummySshKey()
     {
         var keyBits = 2048;
-
         return new SshKeyGenerator.SshKeyGenerator(keyBits);
     }
 }