+++ Frends.SFTP.RenameFile/Frends.SFTP.RenameFile.Tests/RenameTests.cs
@@ -1,6 +1,7 @@
 using NUnit.Framework;
 using System;
 using System.IO;
+using System.Threading.Tasks;
 using Frends.SFTP.RenameFile.Enums;
 
 namespace Frends.SFTP.RenameFile.Tests;
@@ -9,41 +10,41 @@ namespace Frends.SFTP.RenameFile.Tests;
 class RenameTests : RenameFileTestBase
 {
     [Test]
-    public void RenameFile_SimpleTest()
+    public async Task RenameFile_SimpleTest()
     {
-        var result = SFTP.RenameFile(_input, _connection);
+        var result = await SFTP.RenameFile(_input, _connection, default);
         Assert.IsTrue(Helpers.DestinationFileExists(result.Path));
     }
 
     [Test]
-    public void RenameFile_TestRenameBehaviourRename()
+    public async Task RenameFile_TestRenameBehaviourRename()
     {
-        SFTP.RenameFile(_input, _connection);
+        await SFTP.RenameFile(_input, _connection, default);
 
         Helpers.GenerateDummyFile(_input.Path);
         _input.RenameBehaviour = RenameBehaviour.Rename;
-        var result = SFTP.RenameFile(_input, _connection);
+        var result = await SFTP.RenameFile(_input, _connection, default);
         Assert.IsTrue(Helpers.DestinationFileExists(result.Path));
     }
 
     [Test]
-    public void RenameFile_TestRenameBehaviourOverwrite()
+    public async Task RenameFile_TestRenameBehaviourOverwrite()
     {
-        SFTP.RenameFile(_input, _connection);
+        await SFTP.RenameFile(_input, _connection, default);
 
         Helpers.GenerateDummyFile(_input.Path);
         _input.RenameBehaviour = RenameBehaviour.Overwrite;
-        var result = SFTP.RenameFile(_input, _connection);
+        var result = await SFTP.RenameFile(_input, _connection, default);
         Assert.IsTrue(Helpers.DestinationFileExists(result.Path));
     }
 
     [Test]
-    public void RenameFile_TestRenameBehaviourThrow()
+    public async Task RenameFile_TestRenameBehaviourThrow()
     {
-        SFTP.RenameFile(_input, _connection);
+        await SFTP.RenameFile(_input, _connection, default);
 
         Helpers.GenerateDummyFile(_input.Path);
-        var ex = Assert.Throws<ArgumentException>(() => SFTP.RenameFile(_input, _connection));
+        var ex = Assert.ThrowsAsync<ArgumentException>(async () => await SFTP.RenameFile(_input, _connection, default));
         Assert.AreEqual($"File already exists {Path.Combine(Path.GetDirectoryName(_input.Path), _input.NewFileName).Replace("\\", "/")}. No file renamed.", ex.Message);
     }
 }