+++ Frends.SFTP.RenameFile/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [2.0.0] - 2024-01-03
+### Updated
+- [Breaking] Updated dependency SSH.NET to the newest version 2023.0.0.
+
+### Changed
+- Changed connection info builder to create the connection info as it's done in DownloadFiles.
+- [Breaking] Changed PrivateKeyFilePassphrase parameter to PrivateKeyPassphrase and enabled it when PrivateKeyString was used.
+
 ## [1.0.0] - 2023-05-23
 ### Added
 - Initial implementation